# packages/contract/openapi.yaml
openapi: 3.1.0
info:
  title: Loombre API
  version: 1.0.0
  summary: >-
    Contract-first API for Loombre, a ground-up self-hosted media streaming
    platform (movies, TV, music).
  description: >-
    This document is the source of truth (CLAUDE.md architecture invariant
    #1): apps/server controllers conform to it and packages/sdk is generated
    from it — never hand-authored.


    Conventions (apply globally; see components/ for shared shapes): all
    timestamps are BIGINT epoch milliseconds as JSON integers with field
    names suffixed `Ms`; all ids are UUIDv7 strings; all list endpoints use
    cursor pagination (`{ items, nextCursor }`, with `?cursor=&limit=` query
    params); all 4xx/5xx responses are RFC 9457 `application/problem+json`;
    enums are closed OpenAPI enums; core entity schemas set
    `additionalProperties: false`.


    Evolution policy: additive-only within a major version (docs/PLAN.md
    §4.1) — CI runs oasdiff against `main` and fails the build on any
    breaking change classification. Deprecated operations carry
    `deprecated: true` plus a `Sunset` response header for two minor releases
    minimum before a major bump.
  license:
    name: AGPL-3.0-relicense-pending
    identifier: LicenseRef-Loombre-Private
...
servers:
  - url: /v1
    description: Loombre API, version 1 (path-prefixed; host is deployment-specific)
security:
  - bearerAuth: []
...
paths:
...
  /watchlist:
    get:
      operationId: listWatchlist
      tags: [watchlist]
      summary: List the current user's watchlist, newest-added first
      description: >-
        Guarded like every other per-user, item-referencing list (docs/
        PLAN.md §6.4): a row whose item is not visible to the caller right
        now is excluded, even if it was added while the caller was fully
        cleared — restricted titles never appear in the watchlist, locked
        or not (design/phosphor README.md "Restricted content").
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Page of the caller's watchlist
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WatchlistPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Problem' }
  /watchlist/{itemId}:
    parameters:
      - name: itemId
        in: path
        required: true
        schema: { type: string, format: uuid }
    put:
      operationId: addToWatchlist
      tags: [watchlist]
      summary: Add an item to the current user's watchlist (idempotent)
      description: >-
        404 when the item does not exist OR is not visible to the caller —
        indistinguishable, matching every other guarded single-item surface
        (getMovie/getProgress/etc.). This is also what makes adding a
        restricted (zone) item UNREACHABLE without full clearance: the item
        is invisible, not merely disallowed, so there is no signal that
        distinguishes "restricted" from "does not exist" (packages/db/src/
        query/watchlist.ts's addToWatchlistAndEmit).
      responses:
        '204': { description: Added (or already present) }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Problem' }
    delete:
      operationId: removeFromWatchlist
      tags: [watchlist]
      summary: Remove an item from the current user's watchlist (idempotent)
      description: >-
        404 only when the ITEM itself does not exist or is not visible to
        the caller — removing an item that IS visible but was never
        watchlisted is a successful no-op (safe to invoke more than once
        from the inline REMOVE control).
      responses:
        '204': { description: Removed (or was already absent) }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Problem' }

  # ─────────────────────────────── Restricted ───────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 15-minute access JWT; refreshed via per-device rotating refresh tokens.
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque pagination cursor from a previous page's `nextCursor`.
      schema: { type: string }
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
  responses:
    Problem:
      description: Unexpected error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Missing or invalid access token
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    NotFound:
      description: Resource not found (or not visible under the caller's ViewerContext)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 problem details. Extension members are additive.
      properties:
        type: { type: string, format: uri-reference, default: about:blank }
        title: { type: string }
        status: { type: integer, minimum: 100, maximum: 599 }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        code:
          type: string
          description: Stable machine-readable error code, additional to `type`.
      required: [title, status]
      additionalProperties: true
    ItemType:
      type: string
      enum: [movie, series, season, episode, artist, album, track]
    ContentClass:
      type: string
      enum: [general, restricted]
      description: docs/PLAN.md §6.4 — coarse gate on libraries, inherited by items.
...
    ImageKind:
      type: string
      enum: [poster, backdrop, logo, disc, thumb]
...
    ImageDescriptor:
      type: object
      description: >-
        One pre-scaled managed image variant (P1.17). Never computed on
        request (Tier-0 rule) — width/height/blurhash/dominantColor mirror
        the `images` table row an ingest-time worker_thread wrote.
      additionalProperties: false
      required: [kind, width, height, blurhash]
      properties:
        kind: { $ref: '#/components/schemas/ImageKind' }
        width: { type: [integer, 'null'] }
        height: { type: [integer, 'null'] }
        blurhash: { type: [string, 'null'] }
        dominantColor:
          type: [string, 'null']
          description: >-
            Hex '#rrggbb' dominant colour extracted worker-side at ingest
            (P2.11, migrations/0005_images_dominant_color.sql). Null when
            not yet computed/unavailable.

    # ═══════════════════════════════ Auth / devices ═══════════════════════════
...
    CatalogItemBase:
      type: object
      required:
        - id
        - libraryId
        - itemType
        - title
        - sortTitle
        - year
        - communityRating
        - contentClass
        - addedAtMs
        - updatedAtMs
      properties:
        id: { type: string, format: uuid }
        libraryId: { type: string, format: uuid }
        itemType: { $ref: '#/components/schemas/ItemType' }
        title: { type: string }
        sortTitle: { type: string }
        year: { type: [integer, 'null'] }
        communityRating: { type: [number, 'null'], minimum: 0, maximum: 10 }
        contentClass: { $ref: '#/components/schemas/ContentClass' }
        addedAtMs: { type: integer, format: int64 }
        updatedAtMs: { type: integer, format: int64 }

    # Gap-closure lane (deliverable D: "Item detail: metadata, people,
    # versions, play") — mirrors packages/db's item_people/media_files rows.
    # Both PersonRole/PersonCredit and MediaFileSummary are additive: an
    # optional `people`/`mediaFiles` array is added to the relevant item
    # schemas below (movie/series/episode/artist for people; movie/episode/
    # track for mediaFiles), only ever populated by the single-item GET
    # (getMovie/getSeries/... — see packages/db/src/query/catalog-detail.ts's
    # includeDetail param), omitted (not null) on list responses to keep
    # Tier-0 list-page cost unchanged.
    PersonRole:
      type: string
      enum: [actor, director, writer, artist, album_artist, performer, guest]
    PersonCredit:
      description: >-
        Mirrors item_people/people (packages/db/migrations/0001_init.sql):
        `id` is the credited PERSON's id (not the item_people join-row id),
        so a client can link straight to GET /people/{id}. Restricted-class
        people are never credited on a visible general item's response —
        the same content_class join-guard search.ts/catalog-detail.ts
        already use for tags (P1.21) applies here.
      type: object
      additionalProperties: false
      required: [id, name, role, order]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        role: { $ref: '#/components/schemas/PersonRole' }
        credit: { type: [string, 'null'], description: 'e.g. character name ("as Batman").' }
        order: { type: integer, minimum: 0 }
    MediaFileAudioTrack:
      description: >-
        One media_streams audio row on a MediaFileSummary (Phosphor W2 L4
        movie-detail METADATA "Audio" row — packages/db's
        MediaFileAudioTrackSummary).
      type: object
      additionalProperties: false
      required: [codec, channels, language, isDefault]
      properties:
        codec: { $ref: '#/components/schemas/AudioCodec' }
        channels: { type: [integer, 'null'], minimum: 1 }
        language: { type: [string, 'null'], description: ISO 639-2 }
        isDefault: { type: boolean }
    MediaFileSubtitleTrack:
      description: >-
        One media_streams subtitle row on a MediaFileSummary (movie-detail
        METADATA "Subtitles" row — packages/db's
        MediaFileSubtitleTrackSummary).
      type: object
      additionalProperties: false
      required: [language, isForced]
      properties:
        language: { type: [string, 'null'] }
        isForced: { type: boolean }
    MediaFileSummary:
      description: >-
        One media_files row (packages/db/migrations/0001_init.sql), for the
        version/edition picker (multi-version/multi-part items, §8.1) and
        diagnosability. `width`/`height` come from that file's primary video
        stream (media_streams) and are null for audio-only files or files
        not yet probed. `path`/`isDefault`/`videoCodec`/`bitDepth`/`hdr`/
        `audioTracks`/`subtitleTracks` (Phosphor W2 L4 movie-detail VERSIONS
        + METADATA cards) are the same additive, already-probed columns —
        see packages/db's catalog-detail.ts header ("full house" pattern:
        real columns already written by the scanner, newly exposed, nothing
        derived or invented). Deliberately NOT in `required`, unlike the
        original seven fields: this schema doubles as POST /import's
        ExportArchive request body (data-freedom.controller.ts), and an
        archive written by an older Loombre version won't carry them —
        making them required would be a breaking contract change for that
        import path (oasdiff-caught). GET /movies|episodes|tracks/{id}
        always populates every one of them; only a foreign/older import
        payload may omit them.
      type: object
      additionalProperties: false
      required: [id, versionLabel, container, width, height, sizeBytes, durationMs]
      properties:
        id: { type: string, format: uuid }
        versionLabel: { type: [string, 'null'] }
        container:
          type: [string, 'null']
          enum: [mp4, mkv, webm, avi, ts, mov, flac, mp3, ogg, m4a, wav, null]
          description: Null until the file has been probed.
        width: { type: [integer, 'null'] }
        height: { type: [integer, 'null'] }
        sizeBytes: { type: [integer, 'null'], format: int64 }
        durationMs: { type: [integer, 'null'], format: int64 }
        path:
          type: string
          description: >-
            Full on-disk path (media_files.path). Read-only diagnosability
            (docs/PLAN.md §8.4's MediaInfo precedent), same visibility level
            as the rest of this object — never a write target.
        isDefault:
          type: boolean
          description: >-
            True for the item's primary/default file — the unlabelled row
            if one exists, else the earliest-ingested (lowest id) among the
            item's files. Derived per request, not a stored column (no
            multi-version item exists that would make it ambiguous today).
        videoCodec:
          oneOf:
            - $ref: '#/components/schemas/VideoCodec'
            - type: 'null'
          description: Null for audio-only files or files not yet probed.
        bitDepth:
          type: [integer, 'null']
          enum: [8, 10, 12, null]
        hdr:
          oneOf:
            - $ref: '#/components/schemas/HdrType'
            - type: 'null'
          description: Null alongside videoCodec/bitDepth for the same files.
        audioTracks:
          type: array
          items: { $ref: '#/components/schemas/MediaFileAudioTrack' }
        subtitleTracks:
          type: array
          items: { $ref: '#/components/schemas/MediaFileSubtitleTrack' }
    Movie:
      allOf:
        - { $ref: '#/components/schemas/CatalogItemBase' }
        - type: object
          required: [itemType, contentRating, runtimeMs, overview, genres]
          properties:
            itemType: { type: string, const: movie }
            contentRating: { type: [string, 'null'] }
            # Nullable: derived from the media file probe; null until probed.
            runtimeMs: { type: [integer, 'null'], format: int64 }
            overview: { type: [string, 'null'] }
            tagline: { type: [string, 'null'] }
            genres:
              type: array
              items: { type: string }
            images:
              type: array
              items: { $ref: '#/components/schemas/ImageDescriptor' }
              description: >-
                Always present; empty array when no images have been ingested
                yet (Tier-0 rule — never computed on request). Not
                request-required (additive-only: POST /import's ExportArchive
                reuses these same entity schemas as a request body).
            people:
              type: array
              items: { $ref: '#/components/schemas/PersonCredit' }
              description: Only populated by GET /movies/{id}; absent on list responses.
            mediaFiles:
              type: array
              items: { $ref: '#/components/schemas/MediaFileSummary' }
              description: Only populated by GET /movies/{id}; absent on list responses.
      unevaluatedProperties: false
...
    Series:
      allOf:
        - { $ref: '#/components/schemas/CatalogItemBase' }
        - type: object
          required: [itemType, contentRating, overview, status, genres]
          properties:
            itemType: { type: string, const: series }
            contentRating: { type: [string, 'null'] }
            overview: { type: [string, 'null'] }
            # Nullable: from the metadata provider; null until known.
            status: { type: [string, 'null'], enum: [continuing, ended, cancelled, null] }
            genres:
              type: array
              items: { type: string }
            images:
              type: array
              items: { $ref: '#/components/schemas/ImageDescriptor' }
              description: >-
                Always present; empty array when no images have been ingested
                yet (Tier-0 rule — never computed on request). Not
                request-required (additive-only: POST /import's ExportArchive
                reuses these same entity schemas as a request body).
            people:
              type: array
              items: { $ref: '#/components/schemas/PersonCredit' }
              description: Only populated by GET /series/{id}; absent on list responses.
      unevaluatedProperties: false
...
    Album:
      allOf:
        - { $ref: '#/components/schemas/CatalogItemBase' }
        - type: object
          required: [itemType, artistId, genres]
          properties:
            itemType: { type: string, const: album }
            artistId: { type: string, format: uuid }
            trackCount: { type: [integer, 'null'] }
            genres:
              type: array
              items: { type: string }
            images:
              type: array
              items: { $ref: '#/components/schemas/ImageDescriptor' }
              description: >-
                Always present; empty array when no images have been ingested
                yet (Tier-0 rule — never computed on request). Not
                request-required (additive-only: POST /import's ExportArchive
                reuses these same entity schemas as a request body).
      unevaluatedProperties: false
...
    WatchlistEntry:
      type: object
      additionalProperties: false
      required: [itemType, item, addedAtMs]
      properties:
        itemType: { $ref: '#/components/schemas/ItemType' }
        item:
          oneOf:
            - { $ref: '#/components/schemas/Movie' }
            - { $ref: '#/components/schemas/Series' }
            - { $ref: '#/components/schemas/Album' }
          discriminator:
            propertyName: itemType
            mapping:
              movie: '#/components/schemas/Movie'
              series: '#/components/schemas/Series'
              album: '#/components/schemas/Album'
        addedAtMs: { type: integer, format: int64 }
    WatchlistPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/WatchlistEntry' }
        nextCursor: { type: [string, 'null'] }

    # ═══════════════════════════════ Progress ══════════════════════════════════
...
    VideoCodec:
      type: string
      enum: [h264, hevc, av1, vp9, mpeg2, vc1, mpeg4, unknown]
    AudioCodec:
      type: string
      enum: [aac, ac3, eac3, truehd, dts, dtshd, flac, opus, mp3, vorbis, pcm, unknown]
...
    HdrType:
      type: string
      enum: [none, hdr10, hlg, dv]
...